@@ -1,6 +1,6 @@
 {
   "name": "tomato-work",
-  "version": "2.5.0",
+  "version": "3.0.0",
   "homepage": "https://github.com/xjh22222228/tomato-work",
   "bugs": {
     "url": "https://github.com/xjh22222228/tomato-work/issues"
@@ -18,43 +18,43 @@
     "lint": "eslint --cache --ext .js,.jsx,.tsx,.ts ./src"
   },
   "dependencies": {
-    "@ant-design/icons": "^4.6.4",
-    "@toast-ui/editor": "^3.0.3",
-    "antd": "^4.16.13",
-    "axios": "^0.21.1",
-    "blueimp-md5": "^2.18.0",
-    "bytes": "^3.1.0",
+    "@ant-design/icons": "^4.7.0",
+    "@toast-ui/editor": "^3.1.2",
+    "antd": "^4.18.4",
+    "axios": "^0.25.0",
+    "blueimp-md5": "^2.19.0",
+    "bytes": "^3.1.1",
     "classnames": "^2.3.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
-    "query-string": "^7.0.1",
+    "query-string": "^7.1.0",
     "react": "^17.0.2",
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.1",
-    "recharts": "^2.1.2",
-    "redux": "^4.1.1",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.39.0",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.1",
+    "recharts": "^2.1.8",
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.1",
+    "sass": "^1.49.0",
     "use-keep-state": "^1.3.0"
   },
   "devDependencies": {
     "@types/blueimp-md5": "^2.18.0",
     "@types/bytes": "^3.1.1",
-    "@types/lodash": "^4.14.172",
-    "@types/node": "^16.7.10",
-    "@types/react": "^17.0.19",
+    "@types/lodash": "^4.14.178",
+    "@types/node": "^17.0.10",
+    "@types/react": "^17.0.38",
     "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.18",
-    "@types/react-router-dom": "^5.1.8",
-    "@typescript-eslint/eslint-plugin": "^4.30.0",
-    "@typescript-eslint/parser": "^4.30.0",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-redux": "^7.1.22",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "eslint": "^7.32.0",
-    "eslint-plugin-react": "7.25.1",
-    "typescript": "^4.4.2",
-    "vite": "^2.5.3"
+    "eslint": "^8.7.0",
+    "eslint-plugin-react": "7.28.0",
+    "typescript": "^4.5.4",
+    "vite": "^2.7.13"
   }
 }
