@@ -1,22 +1,18 @@
 import React, { FC, useEffect, useState } from 'react'
 import './style.scss'
-import { match, Link, useHistory } from 'react-router-dom'
+import { useParams, Link, useNavigate } from 'react-router-dom'
 import { serviceGetMemorandumById } from '@/services'
 import { defaultTitle } from './constants'
 import { LeftOutlined, EditOutlined } from '@ant-design/icons'
 import { Spin } from 'antd'
 import config from '@/config'
 
-interface Props {
-  computedMatch: match<Record<string, any>>
-}
-
-const DetailPage: FC<Props> = ({ computedMatch }) => {
-  const history = useHistory()
+const DetailPage: FC = () => {
+  const navigate = useNavigate()
   const [title, setTitle] = useState('')
   const [content, setContent] = useState('')
   const [loading, setLoading] = useState(true)
-  const id = computedMatch.params.id
+  const { id } = useParams()
 
   useEffect(() => {
     if (!id) return
@@ -31,11 +27,15 @@ const DetailPage: FC<Props> = ({ computedMatch }) => {
     .finally(() => setLoading(false))
   }, [id])
 
+  function goBack() {
+    navigate('/home/memorandum', { replace: true })
+  }
+
   return (
     <Spin spinning={loading}>
       <div className="memorandum-detail">
         <div className="tool-bar">
-          <LeftOutlined className="icon-left" onClick={history.goBack} />
+          <LeftOutlined className="icon-left" onClick={goBack} />
           <Link className="edit" to={`/home/memorandum/update/${id}`}>
             <EditOutlined title="编辑" />
           </Link>
