@@ -1,6 +1,4 @@
-import config from '@/config'
 import moment from 'moment'
-import { isPlainObject } from 'lodash'
 import { USER } from '../constants'
 import { LOCAL_STORAGE } from '@/constants'
 import { serviceLoginByToken, serviceLogout } from '@/services'
@@ -8,7 +6,7 @@ import { Dispatch } from 'redux'
 
 const { LOGIN } = USER
 
-export function setUser(userInfo: any = {}) {
+export function setUser(userInfo: Record<string, any> = {}) {
   if (userInfo.createdAt) {
     userInfo.createdAt = moment(userInfo.createdAt).format('YYYY-MM-DD')
   }
@@ -35,7 +33,8 @@ export function loginByToken(token: string) {
  */
 export function logout() {
   serviceLogout()
-  .finally(() => {
+
+  setTimeout(() => {
     const localStorageWhiteList = [LOCAL_STORAGE.LOGIN_NAME]
     const localStorageLen = window.localStorage.length
     const allLocalStorageKey: string[] = []
@@ -53,26 +52,4 @@ export function logout() {
     window.sessionStorage.clear()
     window.location.reload()
   })
-}
-
-/**
- * Github Auth
- */
-export function githubAuthz() {
-  const url = `https://github.com/login/oauth/authorize?response_type=code&redirect_uri=${config.github.callbackURL}&client_id=${config.github.clientId}&scope=repo%20repo_deployment%20read:user`
-  window.location.replace(url)
-}
-
-/**
- * 验证本地登录状态
- */
-export function validateLocalStatus() {
-  let userInfo = {}
-  try {
-    userInfo = JSON.parse(window.localStorage.getItem(LOCAL_STORAGE.USER) as string)
-    if (!isPlainObject(userInfo)) {
-      userInfo = {}
-    }
-  } catch {}
-  return setUser(userInfo)
 }
